add icons in ClientManagementPage
