+++ client/src/ui-components/ClientInfoList/ClientInfoListItem.js
@@ -1,5 +1,6 @@
 import { Col, Container, Row } from "react-bootstrap";
 import { FaArrowRight } from "react-icons/fa";
+import { emailIcon, personIcon, phoneIcon, pinMapSmallIcon } from "../icons";
 import "./ClientInfoListItem.css";
 
 function ClientInfoListItem(props) {
@@ -7,10 +8,58 @@ function ClientInfoListItem(props) {
     <Container className="client-info-list-item px-4" onClick={props.onClick}>
       <Row className="py-2">
         <Col className="info px-4">
-          <Row className="client-name mb-2">{props.clientInfo.fullName}</Row>
-          <Row>{props.clientInfo.email}</Row>
-          <Row>{props.clientInfo.phoneNumber}</Row>
-          <Row>{props.clientInfo.address}</Row>
+          <Row className="client-name mb-2">
+            <Col
+              className="px-0"
+              xl="auto"
+              lg="auto"
+              md="auto"
+              sm="auto"
+              xs="auto"
+            >
+              {personIcon}
+            </Col>
+            <Col>{props.clientInfo.fullName}</Col>
+          </Row>
+          <Row>
+            <Col
+              className="px-0"
+              xl="auto"
+              lg="auto"
+              md="auto"
+              sm="auto"
+              xs="auto"
+            >
+              {emailIcon}
+            </Col>
+            <Col>{props.clientInfo.email}</Col>
+          </Row>
+          <Row>
+            <Col
+              className="px-0"
+              xl="auto"
+              lg="auto"
+              md="auto"
+              sm="auto"
+              xs="auto"
+            >
+              {phoneIcon}
+            </Col>
+            <Col>{props.clientInfo.phoneNumber}</Col>
+          </Row>
+          <Row>
+            <Col
+              className="px-0"
+              xl="auto"
+              lg="auto"
+              md="auto"
+              sm="auto"
+              xs="auto"
+            >
+              {pinMapSmallIcon}
+            </Col>
+            <Col>{props.clientInfo.address}</Col>
+          </Row>
         </Col>
         <Col
           className="client-info-list-item-arrow d-flex align-items-center"
